Add separate implementation header for BtActionServer
