+++ nav2_behavior_tree/include/nav2_behavior_tree/bt_action_server_impl.hpp
@@ -1,0 +1,251 @@
+// Copyright (c) 2020 Sarthak Mittal
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef NAV2_BEHAVIOR_TREE__BT_ACTION_SERVER_IMPL_HPP_
+#define NAV2_BEHAVIOR_TREE__BT_ACTION_SERVER_IMPL_HPP_
+
+#include <memory>
+#include <string>
+#include <fstream>
+#include <set>
+#include <exception>
+
+#include "nav2_msgs/action/navigate_to_pose.hpp"
+
+#include "nav2_behavior_tree/bt_action_server.hpp"
+
+namespace nav2_behavior_tree
+{
+
+template<class ActionT>
+BtActionServer<ActionT>::BtActionServer(
+  const rclcpp_lifecycle::LifecycleNode::WeakPtr & parent,
+  const std::string & action_name,
+  OnGoalReceivedCallback on_goal_received_callback,
+  OnLoopCallback on_loop_callback,
+  OnPreemptCallback on_preempt_callback)
+: action_name_(action_name),
+  node_(parent),
+  on_goal_received_callback_(on_goal_received_callback),
+  on_loop_callback_(on_loop_callback),
+  on_preempt_callback_(on_preempt_callback)
+{
+  auto node = node_.lock();
+  logger_ = node->get_logger();
+  clock_ = node->get_clock();
+
+  // Declare this node's parameters
+  node->declare_parameter("default_bt_xml_filename");
+  node->declare_parameter("enable_groot_monitoring", true);
+  node->declare_parameter("groot_zmq_publisher_port", 1666);
+  node->declare_parameter("groot_zmq_server_port", 1667);
+}
+
+template<class ActionT>
+BtActionServer<ActionT>::~BtActionServer()
+{}
+
+template<class ActionT>
+bool BtActionServer<ActionT>::on_configure()
+{
+  auto node = node_.lock();
+  if (!node) {
+    throw std::runtime_error{"Failed to lock node"};
+  }
+
+  // use suffix '_rclcpp_node' to keep parameter file consistency #1773
+  auto options = rclcpp::NodeOptions().arguments(
+    {"--ros-args",
+      "-r", std::string("__node:=") + node->get_name() + "_rclcpp_node",
+      "--"});
+  // Support for handling the topic-based goal pose from rviz
+  client_node_ = std::make_shared<rclcpp::Node>("_", options);
+
+  action_server_ = std::make_shared<ActionServer>(
+    node->get_node_base_interface(),
+    node->get_node_clock_interface(),
+    node->get_node_logging_interface(),
+    node->get_node_waitables_interface(),
+    action_name_, std::bind(&BtActionServer<ActionT>::executeCallback, this));
+
+  // Get the libraries to pull plugins from
+  if (node->has_parameter("plugin_lib_names")) {
+    plugin_lib_names_ = node->get_parameter("plugin_lib_names").as_string_array();
+  } else {
+    RCLCPP_ERROR(logger_, "'plugin_lib_names' param is not defined");
+    return false;
+  }
+
+  // Create the class that registers our custom nodes and executes the BT
+  bt_ = std::make_unique<nav2_behavior_tree::BehaviorTreeEngine>(plugin_lib_names_);
+
+  // Create the blackboard that will be shared by all of the nodes in the tree
+  blackboard_ = BT::Blackboard::create();
+
+  // Put items on the blackboard
+  blackboard_->set<rclcpp::Node::SharedPtr>("node", client_node_);  // NOLINT
+  blackboard_->set<std::chrono::milliseconds>("server_timeout", std::chrono::milliseconds(10));  // NOLINT
+
+  // Get the BT filename to use from the node parameter
+  node->get_parameter("default_bt_xml_filename", default_bt_xml_filename_);
+
+  // Get parameter for monitoring with Groot via ZMQ Publisher
+  node->get_parameter("enable_groot_monitoring", enable_groot_monitoring_);
+  node->get_parameter("groot_zmq_publisher_port", groot_zmq_publisher_port_);
+  node->get_parameter("groot_zmq_server_port", groot_zmq_server_port_);
+
+  return true;
+}
+
+template<class ActionT>
+bool BtActionServer<ActionT>::on_activate()
+{
+  if (!loadBehaviorTree(default_bt_xml_filename_)) {
+    RCLCPP_ERROR(logger_, "Error loading XML file: %s", default_bt_xml_filename_.c_str());
+    return false;
+  }
+  action_server_->activate();
+  return true;
+}
+
+template<class ActionT>
+bool BtActionServer<ActionT>::on_deactivate()
+{
+  action_server_->deactivate();
+  return true;
+}
+
+template<class ActionT>
+bool BtActionServer<ActionT>::on_cleanup()
+{
+  client_node_.reset();
+  action_server_.reset();
+  topic_logger_.reset();
+  plugin_lib_names_.clear();
+  current_bt_xml_filename_.clear();
+  blackboard_.reset();
+  bt_->haltAllActions(tree_.rootNode());
+  bt_->resetGrootMonitor();
+  bt_.reset();
+  return true;
+}
+
+template<class ActionT>
+bool BtActionServer<ActionT>::on_shutdown()
+{
+  return true;
+}
+
+template<class ActionT>
+bool BtActionServer<ActionT>::loadBehaviorTree(const std::string & bt_xml_filename)
+{
+  // Empty filename is default for backward compatibility
+  auto filename = bt_xml_filename.empty() ? default_bt_xml_filename_ : bt_xml_filename;
+
+  // Use previous BT if it is the existing one
+  if (current_bt_xml_filename_ == filename) {
+    RCLCPP_DEBUG(logger_, "BT will not be reloaded as the given xml is already loaded");
+    return true;
+  }
+
+  // if a new tree is created, than the ZMQ Publisher must be destroyed
+  bt_->resetGrootMonitor();
+
+  // Read the input BT XML from the specified file into a string
+  std::ifstream xml_file(filename);
+
+  if (!xml_file.good()) {
+    RCLCPP_ERROR(logger_, "Couldn't open input XML file: %s", filename.c_str());
+    return false;
+  }
+
+  auto xml_string = std::string(
+    std::istreambuf_iterator<char>(xml_file),
+    std::istreambuf_iterator<char>());
+
+  // Create the Behavior Tree from the XML input
+  tree_ = bt_->createTreeFromText(xml_string, blackboard_);
+  topic_logger_ = std::make_unique<RosTopicLogger>(client_node_, tree_);
+
+  current_bt_xml_filename_ = filename;
+
+  // Enable monitoring with Groot
+  if (enable_groot_monitoring_) {
+    // optionally add max_msg_per_second = 25 (default) here
+    try {
+      bt_->addGrootMonitoring(&tree_, groot_zmq_publisher_port_, groot_zmq_server_port_);
+    } catch (const std::logic_error & e) {
+      RCLCPP_ERROR(logger_, "ZMQ already enabled, Error: %s", e.what());
+    }
+  }
+
+  return true;
+}
+
+template<class ActionT>
+void BtActionServer<ActionT>::executeCallback()
+{
+  if (!on_goal_received_callback_(action_server_->get_current_goal())) {
+    action_server_->terminate_current();
+    return;
+  }
+
+  auto is_canceling = [this]() {
+      if (action_server_ == nullptr) {
+        RCLCPP_DEBUG(logger_, "Action server unavailable. Canceling.");
+        return true;
+      }
+      if (!action_server_->is_server_active()) {
+        RCLCPP_DEBUG(logger_, "Action server is inactive. Canceling.");
+        return true;
+      }
+      return action_server_->is_cancel_requested();
+    };
+
+  auto on_loop = [&]() {
+      if (action_server_->is_preempt_requested() && on_preempt_callback_) {
+        on_preempt_callback_();
+      }
+      topic_logger_->flush();
+      on_loop_callback_();
+    };
+
+  // Execute the BT that was previously created in the configure step
+  nav2_behavior_tree::BtStatus rc = bt_->run(&tree_, on_loop, is_canceling);
+
+  // Make sure that the Bt is not in a running state from a previous execution
+  // note: if all the ControlNodes are implemented correctly, this is not needed.
+  bt_->haltAllActions(tree_.rootNode());
+
+  switch (rc) {
+    case nav2_behavior_tree::BtStatus::SUCCEEDED:
+      RCLCPP_INFO(logger_, "Goal succeeded");
+      action_server_->succeeded_current();
+      break;
+
+    case nav2_behavior_tree::BtStatus::FAILED:
+      RCLCPP_ERROR(logger_, "Goal failed");
+      action_server_->terminate_current();
+      break;
+
+    case nav2_behavior_tree::BtStatus::CANCELED:
+      RCLCPP_INFO(logger_, "Goal canceled");
+      action_server_->terminate_all();
+      break;
+  }
+}
+
+}  // namespace nav2_behavior_tree
+
+#endif  // NAV2_BEHAVIOR_TREE__BT_ACTION_SERVER_IMPL_HPP_
